Rename logo.py to plotly.py
