+++ ubicenter/__init__.py
@@ -1,4 +1,4 @@
-from .logo import add_ubi_center_logo, format_fig
+from .plotly import add_ubi_center_logo, format_fig
 
 name = "ubicenter"
 __version__ = "0.1.0"
